refactor: centralize platform path handling
